backend/expense/urls.py
# expenses/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import AccountViewSet

app_name = 'expense'

router = DefaultRouter()
router.register(r'accounts', AccountViewSet, basename='accounts')

urlpatterns = [
    path('expense/', include(router.urls)),
]

